(back) adding validations on the controller, (front) fixing a bug with edit button event

+++ back/src/main/java/co/com/sofka/crud/TodoController.java
@@ -3,6 +3,8 @@ package co.com.sofka.crud;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+
 @RestController
 @CrossOrigin(origins = "http://localhost:3000")
 public class TodoController {
@@ -16,12 +18,15 @@ public class TodoController {
     }
     
     @PostMapping(value = "api/todo")
-    public Todo save(@RequestBody Todo todo){
-        return service.save(todo);
+    public Todo save(@Valid @RequestBody Todo todo){
+        if(todo.getId() != null) {
+            return service.save(todo);
+        }
+        throw new RuntimeException("No existe el id para guardar");
     }
 
     @PutMapping(value = "api/todo")
-    public Todo update(@RequestBody Todo todo){
+    public Todo update(@Valid @RequestBody Todo todo){
         if(todo.getId() != null){
             return service.save(todo);
         }
@@ -29,13 +34,19 @@ public class TodoController {
     }
 
     @DeleteMapping(value = "api/{id}/todo")
-    public void delete(@PathVariable("id")Long id){
-        service.delete(id);
+    public void delete(@PathVariable("id") Long id){
+        if(id != null) {
+            service.delete(id);
+        }
+        throw new RuntimeException("No existe el id");
     }
 
     @GetMapping(value = "api/{id}/todo")
     public Todo get(@PathVariable("id") Long id){
-        return service.get(id);
+        if(id != null) {
+            return service.get(id);
+        }
+        throw new RuntimeException("No existe el id");
     }
 
 }
